Street data can now be parsed.
It still needs an agreement how to represent results.

+++ src/org/mapsforge/applications/debug/Serializer.java
@@ -58,7 +58,7 @@ public class Serializer {
 		return sb.toString();
 	}
 
-	public int getNextShort() {
+	public short getNextShort() {
 		return this.buffer.getShort();
 	}
